+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 5.0.0 (YYYY-MM-DD)
+
+### Breaking Changes
+
+* The `OrderedCollectionChangeSet` parameter in `OrderedRealmCollectionChangeListener.onChange()` is no longer nullable. Use `changeSet.getState()` instead (#5619).
+
+### Enhancements
+
+* Added support for partial Realms. Read [here](https://realm.io/docs/java/latest/#partial-realms) for more information.
+* Added two new methods to `OrderedCollectionChangeSet`: `getState()` and `getError()` (#5619).
+
+### Internal
+
+* Upgraded to Realm Sync 2.2.2.
+
+
 ## 4.3.2 (YYYY-MM-DD)
 
 ### Bug Fixes